Capi-clock pod is running
- Manually generated fake opi certs, which are in capi-values.yaml and referenced in the
opi-secrets configuration. We added a comment specifying that the fake certs are for
capi-k8s-release development purposes only. We need these certs since the clock always
attempts to make a certified connection to eirini. We are getting an expected SocketError
as we don't have a working Eirini component yet in our environment.
-We also made s ome small changes to the deploy.sh script for the dockerfiles and /src/
folder to be referened relative to the repo base directory to allow developers to run the
script from any subdirectory in capi-k8s-release.

[finishes #169924708]

Co-authored-by: Piyali Banerjee <pbanerjee@pivotal.io>
Co-authored-by: Chris Tarazi <ctarazi@pivotal.io>
Co-authored-by: Connor Braa <cbraa@pivotal.io>