Add minimal kpack resource watcher
Right now, this is just a skeleton to be filled in later. This watcher
was generated by kubebuilder and potentially contains useless
boilerplate. There are many things that need to be changed such as the
domain. The intention is to clean this up in the future.

This watcher is called kpack-build-watcher, but it currently wathces
kpack Images.

Co-authored-by: Seth Boyles <sboyles@pivotal.io>
Co-authored-by: Chris Tarazi <ctarazi@pivotal.io>