Remove boilerplate from kubebuilder
We don't need to use kubebuilder because we are not creating our own
CRDs. We only want to hook into kpack's CRDs. We use their Informer
object for instantiation.

Currently, this code only works if you have a KUBECONFIG environment
variable defined, pointing to a kubeconfig file.

This commit squashes:

  - Rename kpack-build-watcher -> capi-kpack-watcher

    This name is more accurate as we add watchers for multiple resources,
    not just kpack Build.

  - Add ability to see updated kpack Builds

Co-authored-by: Chris Tarazi <ctarazi@pivotal.io>
Co-authored-by: Seth Boyles <sboyles@pivotal.io>

[#169373200]