Add kpack-watcher deployment
The kpack-watcher deployment hard codes the CAPI_HOST environment
variable which gives the watcher the URI to talk to CAPI.

Because our Chart is currently for minikube use, deploying this to test
kpack integration won't work. You will need to change the CAPI_HOST
variable to point to katniss.capi.land, for example. In addition, you
will also need to change the tag on the kpack-watcher Dockerfile to
point to a publicly reachable registry.

The kpack-watcher-deployment includes creating ClusterRole and
ClusterRoleBinding so that we can get access to kpack resources.

[#169373200]

Co-authored-by: Chris Tarazi <ctarazi@pivotal.io>
Co-authored-by: Seth Boyles <sboyles@pivotal.io>