Use k8s 1.15.0 version of client-go
The version of client-go must match (or be within the compatibility
matrix[1]) the Kubernetes server version. In our case, katniss runs
Kubernetes 1.14.7.

However, client-go complains about being incompatible with api and
apimachinery at 1.14.7. The error during compilation of the watcher is:

```
\# k8s.io/client-go/rest
/go/pkg/mod/k8s.io/client-go@v11.0.1-0.20190918222721-c0e3722d5cf0+incompatible/rest/request.go:598:31: not enough arguments in call to watch.NewStreamWatcher
        have (*versioned.Decoder)
        want (watch.Decoder, watch.Reporter)
```

Therefore, I chose the next version bump at 1.15.0 which does not have a
compilation problem.

[1]:
https://github.com/kubernetes/client-go/tree/1132e72f6cdf991b5b3f79cac52dbf580639178c#versioning

[#169373200]