Implement an in-cluster Kubernetes client
The build watcher now contains a Kubernetes client to talk to the master
node in order to retrieve pod logs. We can use these pod logs to send
back more detailed error messages to CAPI.

Currently, sending the pod logs to CAPI causes a 500 as a response. We
need to investigate further. This commit brings the scaffolding to order
to do that.

[#169373200]