Make blobstore provider configurable
This PR adds the ability to use GCP, AWS, or Minio by
specifying the provider in values.yml

Changed the default `region` to be `~` (nil) as setting it
as an empty string can cause it to escape the quotes each time
its processed. for example its turned into `''''''` by the time its
passed through the fog lib.

the new fog lib will convert the nil char to an empty string as required.

Signed-off-by: Paul Czarkowski <username.taken@gmail.com>