Re-enqueue build events when they fail without a reason
We've observed that this happens occasionally but haven't tracked down
exactly why. In the cases we've seen, the "Successful" state is "False",
but there are no failure reasons on the containers. Instead, the message
is something like `pods "POD_NAME" already exists`.

See [this kpack issue](https://github.com/pivotal/kpack/issues/460) for context.

In the mean time, we've decided to assume that re-enqueing the event may
address the problem. Our hope is that a later reconciler loop will
encounter a build with a successful state or a failed state with a clear
reason.

If this solution doesn't address the problem, we'll at least have
clearer logging that it is occurring (since the log line will appear on
every attempt to handle the event).

We discussed several strategies to test the re-enqueueing, but couldn't
find one that allowed us to test drive the change. Given that, we've
decided to commit the change untested, with a comment explaining that
fact.

[#173993542]

Co-authored-by: Matt Royal <mroyal@pivotal.io>
Co-authored-by: Jaskanwal Pawar <jpawar@pivotal.io>
Co-authored-by: Nat Bennett <nbennett@pivotal.io>